[Core] Add additional file_exists (errors in the Files class are no longer suppressed)

+++ core-bundle/contao/library/Contao/Files.php
@@ -79,6 +79,11 @@ class Files
 	{
 		$this->validate($strDirectory);
 
+		if (file_exists(TL_ROOT . '/' . $strDirectory))
+		{
+			return true;
+		}
+
 		return mkdir(TL_ROOT . '/' . $strDirectory);
 	}
 
@@ -93,6 +98,11 @@ class Files
 	public function rmdir($strDirectory)
 	{
 		$this->validate($strDirectory);
+
+		if (!file_exists(TL_ROOT . '/' . $strDirectory))
+		{
+			return true;
+		}
 
 		return rmdir(TL_ROOT. '/' . $strDirectory);
 	}
